rename propseditormenu -> templateeditor to better reflect component usage
